test: refactor aws/rke2 test away from caapf
