+++ examples/applications/ccm/aws/helm-chart.yaml
@@ -56,5 +56,4 @@ spec:
           - key: clusterclass-name.fleet.addons.cluster.x-k8s.io
             operator: In
             values:
-              - aws-rke2-example
               - aws-kubeadm-example
