+++ examples/applications/csi/aws/helm-chart.yaml
@@ -17,4 +17,3 @@ spec:
         operator: In
         values:
         - aws-kubeadm-example
-        - aws-rke2-example
